enemy implementation
implemented enemy data mechanics in "game" script

clickMultiplier is now the number of damage the enemy will take
changed playerdata.click from int to float
playerdata.clickTotal -> float

// CONFUSED ABOUT DISPLAYING STUFF, so just used these 2 in "game" script
TextMeshProUGUI enemyHealthDisplay
TextMeshProUGUI levelDisplay

game will display current level and current health of the enemy
vibes will be rewarded upon killing an enemy, based on current level

normal clicks and crit clicks will reduce enemy health instead of giving vibes
when health reaches zero, new enemy image will appear
commented out image changing for crits

quicker crit text animation

+++ Assets/Scripts/PlayerData.cs
@@ -9,8 +9,8 @@ using System;
 public class PlayerData : MonoBehaviour
 {
     // Player Data
-    [SerializeField] public int click = 0;
-    [SerializeField] public int clickTotal = 0;
+    [SerializeField] public float click = 0;
+    [SerializeField] public float clickTotal = 0;
     [SerializeField] public int clickMultiplier = 1;
     [SerializeField] public float critChance = 1f;
     [SerializeField] public int critMultiplier = 3;
